# src/AterStudio/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ater Studio",
    "description": "API 文档",
    "version": "v1"
  },
  "paths": {
...
    "/api/Project": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "获取解决方案列表",
        "operationId": "Project_List",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "添加项目",
        "operationId": "Project_Add",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Project/{id}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "详情",
        "operationId": "Project_Project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Project"
        ],
        "summary": "删除项目",
        "operationId": "Project_Delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Project/sub/{id}": {
      "get": {
        "tags": [
          "Project"
        ],
        "operationId": "Project_GetAllProjectInfos",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubProjectInfo"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubProjectInfo"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubProjectInfo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Project/service": {
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "添加微服务",
        "operationId": "Project_AddService",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Project/setting": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "获取项目配置文件内容",
        "operationId": "Project_GetConfigOptions",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOptions"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOptions"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigOptions"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Project"
        ],
        "summary": "更新配置",
        "operationId": "Project_UpdateConfig",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigOptionsDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigOptionsDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConfigOptionsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Project/open": {
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "打开解决方案，仅支持sln",
        "operationId": "Project_OpenSolution",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Project/tempaltes/{id}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "获取模板名称",
        "operationId": "Project_GetTemplateFiles",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateFile"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateFile"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TemplateFile"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Project/template/{id}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "获取模板内容",
        "operationId": "Project_GetTemplateFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFile"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFile"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TemplateFile"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Project"
        ],
        "summary": "更新模板内容",
        "operationId": "Project_SaveTemplateFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFileUpsert"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFileUpsert"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TemplateFileUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/Project/database/{id}": {
      "get": {
        "tags": [
          "Project"
        ],
        "summary": "获取实体表结构",
        "operationId": "Project_GetDatabaseContent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiDocInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "文档名称"
          },
          "description": {
            "maxLength": 1000,
            "type": "string",
            "description": "文档描述",
            "nullable": true
          },
          "path": {
            "maxLength": 300,
            "type": "string",
            "description": "文档地址"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "接口文档"
      },
...
      "ConfigData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "maxLength": 100,
            "type": "string",
            "description": "键"
          },
          "value": {
            "maxLength": 2000,
            "type": "string",
            "description": "值"
          },
          "valueType": {
            "$ref": "#/components/schemas/ValueType"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "配置"
      },
      "ConfigOptions": {
        "type": "object",
        "properties": {
          "rootPath": {
            "type": "string",
            "description": "项目根目录"
          },
          "isLight": {
            "type": "boolean",
            "description": "是否为轻量项目"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "dtoPath": {
            "type": "string",
            "description": "dto项目目录"
          },
          "entityPath": {
            "type": "string"
          },
          "dbContextPath": {
            "type": "string"
          },
          "applicationPath": {
            "type": "string"
          },
          "apiPath": {
            "type": "string"
          },
          "idStyle": {
            "type": "string",
            "description": "NameId/Id"
          },
          "idType": {
            "type": "string"
          },
          "createdTimeName": {
            "type": "string"
          },
          "updatedTimeName": {
            "type": "string"
          },
          "isSplitController": {
            "type": "boolean",
            "description": "控制器是否拆分",
            "nullable": true
          },
          "controllerType": {
            "$ref": "#/components/schemas/ControllerType"
          },
          "version": {
            "type": "string"
          },
          "webAppPath": {
            "type": "string",
            "description": "前端路径",
            "nullable": true
          },
          "solutionType": {
            "$ref": "#/components/schemas/SolutionType"
          }
        },
        "additionalProperties": false,
        "description": "项目配置"
      },
      "ControllerType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Client",
          "Admin",
          "Both"
        ],
        "x-enumData": [
          {
            "name": "Client",
            "value": 0,
            "description": "用户端"
          },
          {
            "name": "Admin",
            "value": 1,
            "description": "管理端"
          },
          {
            "name": "Both",
            "value": 2,
            "description": "用户端和管理端"
          }
        ]
      },
...
      "EntityInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "类名"
          },
          "namespaceName": {
            "maxLength": 100,
            "type": "string",
            "description": "命名空间",
            "nullable": true
          },
          "assemblyName": {
            "maxLength": 100,
            "type": "string",
            "description": "程序集名称",
            "nullable": true
          },
          "comment": {
            "maxLength": 300,
            "type": "string",
            "description": "类注释",
            "nullable": true
          },
          "summary": {
            "maxLength": 100,
            "type": "string",
            "description": "类注释",
            "nullable": true
          },
          "keyType": {
            "$ref": "#/components/schemas/EntityKeyType"
          },
          "isEnum": {
            "type": "boolean",
            "description": "是否为枚举类",
            "nullable": true
          },
          "isList": {
            "type": "boolean"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "propertyInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PropertyInfo"
            },
            "description": "属性"
          }
        },
        "additionalProperties": false,
        "description": "defined entity model"
      },
      "EntityKeyType": {
        "enum": [
          0,
          1,
          2
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Guid",
          "Int",
          "String"
        ],
        "x-enumData": [ ]
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "项目名称"
          },
          "displayName": {
            "maxLength": 100,
            "type": "string",
            "description": "显示名"
          },
          "path": {
            "maxLength": 200,
            "type": "string",
            "description": "路径"
          },
          "version": {
            "maxLength": 20,
            "type": "string",
            "description": "版本",
            "nullable": true
          },
          "solutionType": {
            "$ref": "#/components/schemas/SolutionType"
          },
          "entityInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityInfo"
            }
          },
          "apiDocInfos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiDocInfo"
            }
          },
          "configDatas": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConfigData"
            }
          },
          "templateFiles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TemplateFile"
            }
          }
        },
        "additionalProperties": false,
        "description": "项目"
      },
      "ProjectType": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Web",
          "Console",
          "Lib",
          "Module",
          "WebAPI",
          "GRPC"
        ],
        "x-enumData": [ ]
      },
      "PropertyInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "isList": {
            "type": "boolean",
            "description": "是否是数组"
          },
          "isPublic": {
            "type": "boolean"
          },
          "isNavigation": {
            "type": "boolean",
            "description": "是否为导航属性"
          },
          "isJsonIgnore": {
            "type": "boolean"
          },
          "navigationName": {
            "type": "string",
            "description": "导航属性类名称",
            "nullable": true
          },
          "isComplexType": {
            "type": "boolean"
          },
          "hasMany": {
            "type": "boolean",
            "description": "导航属性的对应关系",
            "nullable": true
          },
          "isEnum": {
            "type": "boolean"
          },
          "hasSet": {
            "type": "boolean",
            "description": "是否包括set方法"
          },
          "attributeText": {
            "type": "string",
            "nullable": true
          },
          "commentXml": {
            "type": "string",
            "description": "xml comment",
            "nullable": true
          },
          "commentSummary": {
            "type": "string",
            "description": "comment summary",
            "nullable": true
          },
          "isRequired": {
            "type": "boolean",
            "description": "是否必须"
          },
          "isNullable": {
            "type": "boolean",
            "description": "可空?"
          },
          "minLength": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "maxLength": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isDecimal": {
            "type": "boolean"
          },
          "suffixContent": {
            "type": "string",
            "description": "尾缀，如#endregion",
            "nullable": true
          },
          "defaultValue": {
            "type": "string",
            "description": "默认值"
          },
          "entityInfo": {
            "$ref": "#/components/schemas/EntityInfo"
          },
          "entityInfoId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "SolutionType": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "DotNet",
          "Node"
        ],
        "x-enumData": [ ]
      },
      "SubProjectInfo": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "projectType": {
            "$ref": "#/components/schemas/ProjectType"
          }
        },
        "additionalProperties": false,
        "description": "项目信息"
      },
      "TemplateFile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 100,
            "type": "string",
            "description": "名称"
          },
          "displayName": {
            "maxLength": 60,
            "type": "string",
            "description": "显示名称",
            "nullable": true
          },
          "content": {
            "maxLength": 10000,
            "type": "string",
            "description": "内容",
            "nullable": true
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "模板内容"
      },
      "TemplateFileUpsert": {
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 60,
            "type": "string",
            "description": "名称"
          },
          "content": {
            "maxLength": 10000,
            "type": "string",
            "description": "内容"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateConfigOptionsDto": {
        "type": "object",
        "properties": {
          "dtoPath": {
            "type": "string",
            "description": "dto项目目录",
            "nullable": true
          },
          "entityPath": {
            "type": "string",
            "nullable": true
          },
          "entityFrameworkPath": {
            "type": "string",
            "nullable": true
          },
          "storePath": {
            "type": "string",
            "nullable": true
          },
          "apiPath": {
            "type": "string",
            "nullable": true
          },
          "idType": {
            "type": "string",
            "nullable": true
          },
          "createdTimeName": {
            "type": "string",
            "nullable": true
          },
          "updatedTimeName": {
            "type": "string",
            "nullable": true
          },
          "controllerType": {
            "$ref": "#/components/schemas/ControllerType"
          },
          "isSplitController": {
            "type": "boolean",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ValueType": {
        "enum": [
          0,
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32",
        "x-enumNames": [
          "Integer",
          "Double",
          "String",
          "Boolean"
        ],
        "x-enumData": [ ]
      }
    }
  },
...
}